기능삭제: 퀴즈 도메인 삭제, 프로젝트명 및 패키지명 변경
