+++ README.md
@@ -4,3 +4,7 @@
 ### 2024.01.09 변경사항
 - 프로젝트 목표 변경: 퀴즈 웹 서비스 → 학생 관리 웹 서비스  
 - 커밋 메세지 컨벤션 변경: 영문 → 한글
+
+### 2024.01.10 변경사항
+- 프로젝트 및 패키지명 리팩토링
+- 프론트엔드 영역 변경: 리액트 삭제, 추후 방향 결정
