--- react-app/src/App.js
@@ -1,12 +1,0 @@
-import './App.css';
-import Login from './components/Login';
-
-function App() {
-  return (
-    <div className="App">
-      <Login />
-    </div>
-  );
-}
-
-export default App;
