+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'quiz'
+rootProject.name = 'student-management-web-service'
