--- src/main/java/kr/co/quiz/QuizApplication.java
@@ -1,12 +1,0 @@
-package kr.co.quiz;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class QuizApplication {
-	public static void main(String[] args) {
-		SpringApplication.run(QuizApplication.class, args);
-	}
-
-}
