--- src/main/java/kr/co/quiz/game/entity/Game.java
@@ -1,35 +1,0 @@
-package kr.co.quiz.game.entity;
-
-import jakarta.persistence.*;
-import kr.co.quiz.user.entity.User;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-@Table(name = "tbl_game")
-public class Game {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "game_id")
-    private Long id;
-    @Column
-    private String title;
-    @Column
-    private String description;
-
-    @ManyToOne
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    @Builder
-    public Game(String title, String description, User user) {
-
-        this.title = title;
-        this.description = description;
-        this.user = user;
-    }
-}
