--- src/main/java/kr/co/quiz/game/entity/GameRepository.java
@@ -1,9 +1,0 @@
-package kr.co.quiz.game.entity;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface GameRepository extends JpaRepository<Game, Long> {
-    Optional<Game> findOptionalById(Long id);
-}
