--- src/main/java/kr/co/quiz/user/dtos/UserReqDTO.java
@@ -1,14 +1,0 @@
-package kr.co.quiz.user.dtos;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-@AllArgsConstructor
-public class UserReqDTO {
-    private String username;
-    private String password;
-
-}
