--- src/main/java/kr/co/quiz/user/entity/Role.java
@@ -1,5 +1,0 @@
-package kr.co.quiz.user.entity;
-
-public enum Role {
-    USER, MANAGER, ADMIN
-}
