--- src/main/java/kr/co/quiz/user/entity/UserRepository.java
@@ -1,9 +1,0 @@
-package kr.co.quiz.user.entity;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findOptionalByUsername(String username);
-}
