+++ src/main/java/kr/co/smws/SmwsApplication.java
@@ -1,0 +1,12 @@
+package kr.co.smws;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SmwsApplication {
+	public static void main(String[] args) {
+		SpringApplication.run(SmwsApplication.class, args);
+	}
+
+}
