+++ src/main/java/kr/co/smws/global/config/EnvProperties.java
@@ -1,0 +1,17 @@
+package kr.co.smws.global.config;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.core.env.Environment;
+import org.springframework.stereotype.Component;
+
+@Component
+@PropertySource("file:C:/tools/quiz-server.txt")
+public class EnvProperties {
+    @Autowired
+    private Environment env;
+
+    public void myMethod() {
+        String SECRET_KEY = env.getProperty("SECRET_KEY");
+    }
+}
