+++ src/main/java/kr/co/smws/global/config/jwt/JwtToken.java
@@ -1,0 +1,14 @@
+package kr.co.smws.global.config.jwt;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+@AllArgsConstructor
+public class JwtToken {
+    private String grantType;
+    private String accessToken;
+    private String refreshToken;
+}
