+++ src/main/java/kr/co/smws/global/config/security/UserDetailsDTO.java
@@ -1,0 +1,56 @@
+package kr.co.smws.global.config.security;
+
+import kr.co.smws.user.entity.User;
+import lombok.Getter;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+
+import java.util.ArrayList;
+import java.util.Collection;
+
+@Getter
+public class UserDetailsDTO implements UserDetails {
+    private User user;
+    public UserDetailsDTO(User user) { this.user = user; }
+
+    @Override
+    public String getPassword() {
+        return user.getPassword();
+    }
+
+    @Override
+    public String getUsername() {
+        return user.getUsername();
+    }
+
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
+
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
+
+    @Override
+    public boolean isEnabled() {
+        return true;
+    }
+
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        Collection<GrantedAuthority> collectors = new ArrayList<>();
+        collectors.add(()->{
+
+            return "ROLE_" + user.getRole();
+        });
+
+        return collectors;
+    }
+}
