+++ src/main/java/kr/co/smws/user/dtos/UserReqDTO.java
@@ -1,0 +1,14 @@
+package kr.co.smws.user.dtos;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+@AllArgsConstructor
+public class UserReqDTO {
+    private String username;
+    private String password;
+
+}
