+++ src/main/java/kr/co/smws/user/entity/Role.java
@@ -1,0 +1,5 @@
+package kr.co.smws.user.entity;
+
+public enum Role {
+    USER, MANAGER, ADMIN
+}
