+++ src/main/java/kr/co/smws/user/entity/UserRepository.java
@@ -1,0 +1,9 @@
+package kr.co.smws.user.entity;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    Optional<User> findOptionalByUsername(String username);
+}
