--- src/test/java/kr/co/quiz/QuizApplicationTests.java
@@ -1,13 +1,0 @@
-package kr.co.quiz;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class QuizApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
