+++ src/test/java/kr/co/smws/SmwsApplicationTests.java
@@ -1,0 +1,13 @@
+package kr.co.smws;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SmwsApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
